refactor(lifecycle): split app_setup.rs into focused modules
Extract skills manifest helpers, managed/assistant skill sync, and
startup orchestration into app_setup/ while preserving the public API
and existing integration test coverage.
